requirements/development.txt
# requirements for development and testing
setuptools==63.1.0
build==0.8.0
python-semantic-release==7.29.4
pytest==7.1.2
black==22.6.0
pytest-dotenv==0.5.2

# superglue requirements
python-dotenv==0.20.0
pyaml==21.10.1
boto3==1.20.49
Jinja2==3.0.3
